apitracking/x/apitracking/types/message_api_request_test.go
package types

import (
	"testing"

	"apitracking/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgApiRequest_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgApiRequest
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgApiRequest{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgApiRequest{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
